Configurable toolbars moved to iiwaConfiguration; available from Monitor and SmartServo
